Improve status reporting and cache resource 'generation'
There are multiple changes in this patch:

Status.phase is set to 'Deploying' at the start
of the reconciliation iteration.

If an error occurs, it is reported in conditions.

When checking if a resource was modified, the operator
compares a cached 'generation' value. If the resource
does not use generation, operator compares the 'resourceVersion'

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>