Break infinite reconciliation loop (#59)
* Break infinite reconciliation loop

Since Openshift 4.6 and Kubernetes v1.19.0,
a status change on SSP CR causes reconciliation.

This patch adds a predicate to ignore the status change,
otherwise the operator would never stop with reconciliation.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>

* Compare available deployment replicas with current required value

Checking if unavailable replicas is not zero is incorrect,
because the observed generation may be different
from current generation, and the status could relate to old
spec.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>

* Improve status reporting and cache resource 'generation'

There are multiple changes in this patch:

Status.phase is set to 'Deploying' at the start
of the reconciliation iteration.

If an error occurs, it is reported in conditions.

When checking if a resource was modified, the operator
compares a cached 'generation' value. If the resource
does not use generation, operator compares the 'resourceVersion'

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>