Fix VERSION variable in Makefile, fix Dockerfile, update CSV
The VERSION variable in makefile should be equal to the next
release version in the sequence.

A file was not copied to the builder image in Dockerfile.
It produces an error message, that is currently ignored, but
when the operator-sdk will be updated, it will cause the build to fail.

Regenerated CSV in data/olm-catalog.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>