Add validating webhook for templates
The webhook is called when a template is deleted.
It checks if the deleted template is an old style VM
template with validation annotation, and if any VMs
are referencing it.

This commit adds VmCache structure that implements cache.Store
interface. It uses a predicate to only store relevant VMs,
and it only stores useful data, not the full VM structure.

This VmCache is used in cache.Reflector to watch all VMs
in the cluster.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>