Detect if a resource is being deleted and wait for cleanup
The reconciliation function did not handle the case
when a resource is being deleted, but is not gone yet.
This is noticeable for namespace filled with other resources.

After this change, resources that are being deleted
are treated as not ready, and the SSP resource will
not be in Deployed state.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>