Add initial controller to watch missing CRDs
If not all needed CRDs are installed, a controller
that watches and waits for CRDs will be started.

Once all needed CRDs are installed, the initial controller
will exit and the SSP controller will start.

These CRDs are needed:
- datavolumes.cdi.kubevirt.io
- datasources.cdi.kubevirt.io
- dataimportcrons.cdi.kubevirt.io
- prometheusrules.monitoring.coreos.com (is installed with OCP)

The SSP controller watches these resources and it would fail,
if CRDs do not exist, and the application will exit with an error.
This would cause the operator deployment to enter a crash loop.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>