fix applying deprecated label
the application of deprecated label was not working correctly,
because it doesn't skip old templates with the same name as newly
deployed templates.
Added a condition, which check if template has newer version than is
the version in ssp operator. If yes, template is not set as deprecated.
If parsing of template version fails, continue with adding deprecated label.

Signed-off-by: Karel Šimon <ksimon@redhat.com>