Remove selected template labels on reconcile
The common reconcile method does not remove labels
and annotations from resources, which is useful
to not remove user-defined labels and annotations.

This patch adds a special logic to remove labels
and annotations from templates, if they are not
user-defined. For example, if the default template
for an OS changes, then the label needs to be removed
from the old template.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>