Deploying state and tests for TLSSecurityProfile
While the pod is restarting during TLSSecurityProfile is update
we will be in a "deploying" state.
Add functional tests for the various profiles.

Conflicts:
  tests/crypto_policy_test.go

NOTE(0xFelix): The conflict is due to not having the tests/env package
in release-0.16. The file tests/crypto_policy_test.go is changed to used
the old timeout variable defined in tests/tests_suite_test.go.

Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>
Signed-off-by: borod108 <boris.od@gmail.com>