Allow running the SSP Operator on plain kubernetes
SSP can't run on a plain kubernetes, because it requires the
`Infrastructure` kind to be present on the cluster.

When deploying KubeVirt with HCO using OLM on plain kubernetes,
the deployment is never completed because SSP operator never becomes
ready. More details and full description of the issue can be found here:
https://github.com/kubevirt/hyperconverged-cluster-operator/issues/2129

This commit disables the reconcilers and all the watches, when running
on a plain kubernetes cluster, instead of killing the process.

That way, the SSP operator
is runing and responding to the health and ready checks, but does
nothing else. The result is that SSP is still not working on plain
kubernetest, but the OLM deployment is successfully completed.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>