dataSources_test.go: Reenable CDI garbage collection
CDI should now be able to properly handle DataImportCrons with garbage
collection enabled. Therefore it is no longer needed to disable the
garbage collection.

The tests are updated to handle the garbage collection of DataVolumes
used to populate DataSources and to delete PVCs of garbage collected
DataVolumes.

  tests/dataSources_test.go

NOTE: The conflicts were due to changes to the timeout variable missing
in this branch. The patch was adapted to use the old timeout variable.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>