crypto_policy_test.go: Do not hardcode SSP webhook service name
Introduce a getter for the SSP webhook service name and use it instead
of hardcoding the name to 'ssp-webhook-service'.

Conflicts:
  tests/env/env.go

NOTE(0xFelix): The conflict is due to not having the tests/env package
in release-0.16. The variable envSspWebhookServiceName is added to
tests/tests_suite_test.go instead and used there.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>