data_sources: Always run the operand
The operand itself deploys DataImportCrons defined by the SSP CR and
DataSources based on the provided common-templates bundle. While the
latter functionality should likely reside in the common-templates
operand the DataImportCrons functionaility is useful in non-OpenShift
environments alongside that of the common-instancetypes operand.

This change now ensures the data_sources operand runs on both OpenShift
and non-OpenShift environments.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>