Add alert for rhel6 running vms
RHEL6 guests will be not supported on RHEL9 hosts anymore [1].
And CNV 4.13 is planned to be on RHCOS 9, so RHEL6 cannot be supported in CNV 4.13 anymore.

Adding a vm controller that watch for rhel 6 vms.
Rhel6 template put a label in the vm with the format:
rhel6-<workload>-<flavor>.
This label is checked to determine whether a vm is rhel6 or not.

[1] https://access.redhat.com/articles/973163#rhelkvm

Signed-off-by: fossedihelm <ffossemo@redhat.com>