common-instancetypes: Reconcile from a remote kustomize target
This change introduces support within the operand for reconciling from a
remote kustomize target. This support mirrors the underlying kustomize
remote target support for https:// and ssh:// URLs. Additionally
the use of either ?ref= or ?version= parameters in the URL is required
to ensure that the contents of the URL do not change overtime.

For now the operand will only reconcile a URL initially on start up of
the SSP manager *or* if the URL changes again during the lifetime of the
manager. Future changes will likely make this more flexible but this is
the behaviour of this initial implementation to keep things simple.

This change has also introduced the removal of removed bundled resources
as the logic is the same when using the bundle or URL approach.

The bulk of the change is caused by the vendoring of the kustomize
krusty entry point module used to render resources from the given URI.

The git package is included in the manager container image and is a
direct dependency of Kustomize when using remote git repositories.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>