fix: functest: Wait after reverting to original SSP CR
Functions RevertToOriginalSspCr() now wait until SSP
is deployed before returning.

This is useful, because reverting SSP may change TLS
policy, which causes the pod to crash, and it may
be some time before it is running again.
If in the meantime, another test tries to update the SSP,
it will fail, because the webhook cannot be reached.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>