feat: rework example pipelines
Common-templates are going to be deprecated in some future version.
This commit is changing example pipelines to remove all automation
related to templates and replaces it with create-vm-from-manifest task.

This change helps to unify pipelines for okd and kubernetes, so now we
can use single version for okd and k8s.

windows-bios-installer pipeline is redesigned to run only in a single
namespace, which helps to reduce number of steps and complexity.

windows-customize pipeline is redesigned to run only in a single
namespace which helps to reduce number of steps and complexity.

Due to changes in pipelines, there were done necessary changes in RBAC -
objects separated in different file, rules defined in bios-installer
pipeline were replaced by roleBindings to original task's clusterRoles.

Signed-off-by: Karel Šimon <ksimon@redhat.com>