fix: Don't replace namespace annotation when kubevirt namespace is set in ssp CR
When ssp CR contains kubevirt namespace as a target namespace
for pipelines, ssp operator recognized this namespace as user defined
and deployed configmaps and roleBindings to a single namespace, instead
of two namespaces. That broke our default pipeline flow. This
change adds a check if the pipeline namespace in the CR is kubevirt,
then it is not used as a user defined namespace and
kubevirt.io/tekton-piplines-deploy-namespace annotation is used instead.

rename kubevirt.io/deploy-namespace annotation to kubevirt.io/tekton-piplines-deploy-namespace
Signed-off-by: Karel Simon <ksimon@redhat.com>