fix: Increase reconcile metric only if SSP CR was not changed
The metric "kubevirt_ssp_common_templates_restored_total" should
be increased if user changes a template and the operator restores
it back to what it should be.

If the SSP CR is changed, for example one of the labels
that are propagated to the templates are changed,
then the reconciliation should not increase the metric.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>