fix: disable SSP_BUILD_RUNTIME and use podman only
The following reasons support this change:
1. The commands for creating multi-architecture images differ
between docker and podman.
2. The docker manifest command requires that images be uploaded to
a registry first, which complicates the build process by merging
the building and pushing steps.
3. While some consistency could be achieved using docker buildx,
the commands would still differ and would require a host capable
of emulating other architectures for Podman users.
4. I did not find any references to the SSP_BUILD_RUNTIME variable
in the release repository: https://github.com/openshift/release

Overall, this change aims to simplify the process.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>