tests: On cleanup, wait until PVCs are bound.
When running functional tests in an existing SSP deployment,
the DataImportCrons are removed and recreated multiple times.
The PVCs that they point to can be removed in the process.

This commit adds code in AfterSuite() to wait for the PVCs
to be imported again.

One of the tests was removing the namespace that contains
DataImportCrons and imported images. This caused a problem
when the same namespace also contained ImageStreams
for the imported images.

This commit removes the test case, because it is already
covered in unit tests.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>