feat: enable multiarch image build by default
To achieve this, we need to implement additional logic in the
controller-gen, alongside the traditional process of building for
different architectures and combining them into a single manifest.
The challenge arises from the fact that GOBIN cannot be used in
conjunction with GOARCH.

While the workaround is somewhat hacky, I have not identified a
more straightforward solution.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>