api: Update SSP API version to v1beta3
The new API version removes feature gates and unused
parameters. It is directly compatible with the old version,
so a conversion webhook is not needed.

The stored version is still set to v1beta2, and ssp-operator
watches this old version. In a next release we may switch
to using the v1beta3 as stored version.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>