tests: TLS policy: Use insecure TLS for connection to template-validator
The current code in functional tests that loads root CA for TLS
connection to template-validator does not work when SSP is deployed by
OLM.

This commit is a simple fix, where we ignore the validity of the
certificate passed by template-validator. This is ok for these tests,
because different tests that check behavior of the template-validator
would fail if the certificate was invalid.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>