chore(test): Add cleanup in VM delete protection test
There is a race condition between the deletion in those test in which
the deletion is expected to happen and the `AfterEach` `IsNotFound` check.
This may cause that the `AfterEach` tries to update the VM to remove the
label but the VM is already gone.

It adds a `WaitForDeletion` in those test to be sure that the VM is removed
before triggering the `AfterEach`. Moreover, it adds a check in the
`AfterEach` to check if the VM is already being deleted.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>