functests: create port-forward connection in DialContext
There are tests that check that invalid TLS connections are rejected.
When a connection is rejected, the client closes it.
This caused an error when the test code tried the close the connection
again.

This PR moves the creation of the connection to the DialContext function.
It is now managed by the HTTP client.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>