fix(instancetypes): Protect instancetype/preference `controllerRevision` objects
This protection ensures that instancetype and preference
`controllerRevision` objects are not deleted before `VirtualMachine`
objects. This is due to the fact that in some scenarios, such as when a
VM is delete protected, if the `controllerRevision` is deleted but the
VM object is still alive, users will not be able to patch or update the
VM object. This leads to virtually deadlocking the VM.

The VAP ensures that only the Kubernetes garbage collector and
kubevirt-controller are able to delete these objects:

* Garbage-collector: It deletes objects when the `ownerReference`'d
  object is removed, e.g., the instancetype "test-it" with
  `ownerReference: test-vm` is removed when the VM test-vm is removed.
  Therefore, by only allowing it to drop these objects, we ensure the
  deletion order.
* Kubevirt-controller: It is required to no block this user because is
  the main owner of these `controllerRevision`, it just deletes them in
  upgrade scenarios, live migration, etc.

Fix: https://issues.redhat.com/browse/CNV-69511

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>