fix: Use PATCH call to update metadata of immutable objects
When a new field is added to the API of an immutable object
before SSP has the new API code imported, the operator will
ignore the new fields.

If it tries to update metadata, the UPDATE request will not
contain the new field in spec, and so it will look like spec
modification. The webhook would reject it.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>